HashMap internal working and Index
